+++ changelog.md
@@ -1,3 +1,9 @@
+# v1.0.3
+
+- added support for item info popup arrows on profile pages, for people with Better Unlock Info
+- changed names of elements from 'Unlock Popup' to 'Item Info Popup' in settings and about pages
+- move arrows in item info popups slightly up, to make them centered on the popup
+
 # v1.0.2
 
 - remove log i forgot to remove